LocalMMScene: put VR Scene in own file
