+++ src/test/kotlin/microscenery/example/microscope/LocalMMScene.kt
@@ -1,16 +1,11 @@
 package microscenery.example.microscope
 
-import graphics.scenery.controls.behaviours.Switch
-import graphics.scenery.controls.behaviours.WheelMenu
-import graphics.scenery.volumes.Volume
 import microscenery.*
 import microscenery.UI.StageSpaceUI
-import microscenery.VRUI.VRUIManager
 import microscenery.hardware.MicroscopeHardware
 import microscenery.hardware.micromanagerConnection.MMConnection
 import microscenery.hardware.micromanagerConnection.MicromanagerWrapper
 import microscenery.signals.ServerState
-import microscenery.stageSpace.FrameGizmo
 import microscenery.stageSpace.MicroscopeLayout
 import microscenery.stageSpace.StageSpaceManager
 import mmcorej.CMMCore
@@ -34,9 +29,9 @@ class LocalMMScene : DefaultScene(withSwingUI = true) {
         MicroscenerySettings.set("Stage.minX", -300f)
         MicroscenerySettings.set("Stage.minY", -300f)
         MicroscenerySettings.set("Stage.minZ", -250f)
-        MicroscenerySettings.set("Stage.maxX", 300f)
-        MicroscenerySettings.set("Stage.maxY", 300f)
-        MicroscenerySettings.set("Stage.maxZ", 250f)
+        MicroscenerySettings.set("Stage.maxX", 3000f)
+        MicroscenerySettings.set("Stage.maxY", 3000f)
+        MicroscenerySettings.set("Stage.maxZ", 2500f)
 
         val stageStart = Vector3f()
 
@@ -101,68 +96,3 @@ class LocalMMScene : DefaultScene(withSwingUI = true) {
 
 }
 
-class LocalMMSceneVR : DefaultVRScene() {
-
-    lateinit var stageSpaceManager: StageSpaceManager
-
-    override fun init() {
-        super.init()
-
-        MicroscenerySettings.set("Stage.minX", -1000f)
-        MicroscenerySettings.set("Stage.minY", -1000f)
-        MicroscenerySettings.set("Stage.minZ", -500f)
-        MicroscenerySettings.set("Stage.maxX", 1000f)
-        MicroscenerySettings.set("Stage.maxY", 1000f)
-        MicroscenerySettings.set("Stage.maxZ", 500f)
-
-        val stageStart = Vector3f()
-
-        val hardware: MicroscopeHardware =
-            MicromanagerWrapper(MMConnection(LocalMMScene.initLocalMMCoreFake()).apply { moveStage(stageStart, false) })
-        stageSpaceManager = StageSpaceManager(
-            hardware, scene, hub, addFocusFrame = true, scaleDownFactor = 50f,
-            layout = MicroscopeLayout.Default(MicroscopeLayout.Axis.Z)
-        )
-    }
-
-    override fun inputSetup() {
-        super.inputSetup()
-        inputHandler?.let {
-            StageSpaceUI(stageSpaceManager).stageKeyUI(it, cam)
-        }
-
-        VRUIManager.initBehavior(
-            scene, hmd, inputHandler,
-            customActions = WheelMenu(
-                hmd, listOf(
-                    Switch("Steering", false) { value ->
-                        stageSpaceManager.focusTarget?.let {
-                            if (value && it.mode != FrameGizmo.Mode.STEERING) {
-                                it.mode = FrameGizmo.Mode.STEERING
-                            } else {
-                                it.mode = FrameGizmo.Mode.PASSIVE
-                            }
-                            logger.info("focusframe mode is now ${it.mode}")
-                        }
-                    },
-                    Switch("Live", false) {
-                        if (it)
-                            stageSpaceManager.goLive()
-                        else
-                            stageSpaceManager.stop()
-                    },
-                )
-            )
-        ) {
-            scene.findByClassname(Volume::class.simpleName!!).first() as Volume
-        }
-    }
-
-    companion object {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            LocalMMSceneVR().main()
-        }
-    }
-
-}
